Put ChecklistDoc into its own class file
